requirements.txt
DrissionPage==4.0.4.21
aiofiles==23.2.1
pillow~=10.2.0